more validations for routes and catching of erros

+++ app/Http/Controllers/GroupController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Group;
 use App\Models\Membership;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class GroupController extends Controller{
     public function __construct() {
@@ -27,21 +28,56 @@ class GroupController extends Controller{
         if($group){
             return response()->json($group, 200);
         }else{
-            return response()->json("Can't get group.", 404);
+            return response()->json("Group doesn't exist.", 404);
         }
     }
 
     public function create(Request $request){
+        $this->validate($request, [
+            'passwort' => 'required',
+            'hinweis' => 'required',
+            'admin_P_ID' => 'required',
+            'name' => 'required|unique:gruppen'
+        ]);
+
         $group = new Group();
 
-        $group->passwort    = $request->passwort;
+        $group->passwort    = Hash::make($request->passwort);
         $group->hinweis     = $request->hinweis;
         $group->admin_P_ID  = $request->admin_P_ID;
+        $group->name        = $request->name;
 
-        $check = $group->save();
+        $checkGroupCreated = $group->save();
 
-        if($check){
-            return response()->json("Group Successfully Created!", 200);
+        if($checkGroupCreated){
+            $group = Group::where('name', '=', $request->name)
+                            ->where('hinweis', '=', $request->hinweis)
+                            ->where('admin_P_ID', '=', $request->admin_P_ID)->first();
+
+            if($group->G_ID){
+                $membership = new Membership();
+
+                $membership->P_ID    = $request->admin_P_ID;
+                $membership->G_ID    = $group->G_ID;
+    
+                $checkMembership = $membership->save();
+
+                if($checkMembership){
+                    return response()->json("Group Successfully Created!", 200);
+                }else{
+                    $group = Group::where('name', '=', $request->name)
+                                    ->where('hinweis', '=', $request->hinweis)
+                                    ->where('admin_P_ID', '=', $request->admin_P_ID)->first();
+                    $group->delete();
+                    return response()->json("Creating the group failed!", 500);
+                }
+            }else{
+                $group = Group::where('name', '=', $request->name)
+                            ->where('hinweis', '=', $request->hinweis)
+                            ->where('admin_P_ID', '=', $request->admin_P_ID)->first();
+                $group->delete();
+                return response()->json("Creating the group failed!", 500);
+            }
         }else{
             return response()->json("Creating the group failed!", 500);
         }
@@ -50,30 +86,49 @@ class GroupController extends Controller{
     public function update(Request $request, $G_ID){
         $group = Group::find($G_ID);
 
-        $group->passwort    = $request->passwort;
-        $group->hinweis     = $request->hinweis;
+        if($group){
+            $this->validate($request, [
+                'passwort' => 'required',
+                'hinweis' => 'required'
+            ]);
 
-        $check = $group->save();
+            $group->passwort    = Hash::make($request->passwort);
+            $group->hinweis     = $request->hinweis;
 
-        if($check){
-            return response()->json($group, 200);
+            $check = $group->save();
+
+            if($check){
+                return response()->json($group, 200);
+            }else{
+                return response()->json('Updating the group failed!', 500);
+            }
         }else{
-            return response()->json('Updating the group failed!', 500);
+            return response()->json("Group doesn't exist.", 404);
         }
     }
 
     public function delete($G_ID){
         $group = Group::find($G_ID);
-        $check = $group->delete();
 
-        if($check){
-            return response()->json('Group Successfully Deleted!', 200);
+        if($group){
+            $check = $group->delete();
+
+            if($check){
+                return response()->json('Group Successfully Deleted!', 200);
+            }else{
+                return response()->json('Deleting the group failed!', 500);
+            }
         }else{
-            return response()->json('Deleting the group failed!', 500);
+            return response()->json("Group doesn't exist.", 404);
         }
     }
 
     public function join(Request $request){
+        $this->validate($request, [
+            'P_ID' => 'required',
+            'G_ID' => 'required'
+        ]);
+
         $membership = new Membership();
 
         $membership->P_ID    = $request->P_ID;
